Denom remainderify

+++ Monkey.cs
@@ -1,13 +1,15 @@
+using System.Numerics;
+
 public class Monkey
 {
 	public int inspected = 0;
-	public List<int> items;
-	public Func<int, int> operation;
-	public Func<int, bool> test;
+	public List<BigInteger> items;
+	public Func<BigInteger, BigInteger> operation;
+	public Func<BigInteger, bool> test;
 	public int trueTarget;
 	public int falseTarget;
 
-	public Monkey(List<int> items, Func<int, int> operation, Func<int, bool> test, int trueTarget, int falseTarget)
+	public Monkey(List<BigInteger> items, Func<BigInteger, BigInteger> operation, Func<BigInteger, bool> test, int trueTarget, int falseTarget)
 	{
 		this.items = items;
 		this.operation = operation;
